032_function, 032_test (pep8, pylint, test)
